Switching from ConfigReader to JText.

+++ Animation.h
@@ -5,15 +5,15 @@ namespace EngineCore {
 	class Animation {
 	private:
 		std::vector<Texture> frames = {};
-		Uint32 current = 0;
+		size_t current = 0;
 
 		clock_t lastTime = 0;
-		Uint32 cycle = 0;
+		size_t cycle = 0;
 	public:
 		Sprite* sprite = nullptr;
 
 		clock_t interval = 100;
-		Uint32 cycles = 1;
+		size_t cycles = 1;
 		bool loop = false;
 
 		void addFrame(Texture frame);
